+++ device_preview/example/lib/main.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:device_preview/device_preview.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
@@ -8,16 +6,8 @@ import 'package:intl/intl.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_platform_widgets/flutter_platform_widgets.dart';
 
-// Sets a platform override for desktop to avoid exceptions. See
-// https://flutter.dev/desktop#target-platform-override for more info.
-void _enablePlatformOverrideForDesktop() {
-  if (!kIsWeb && (Platform.isMacOS || Platform.isWindows || Platform.isLinux)) {
-    debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
-  }
-}
-
 void main() {
-  _enablePlatformOverrideForDesktop();
+  debugDefaultTargetPlatformOverride = TargetPlatform.fuchsia;
   runApp(DevicePreview(
     builder: (context) => MyApp(),
   ));
@@ -26,19 +16,14 @@ void main() {
 class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    final mediaQuery = MediaQuery.of(context);
+    final mediaQuery = MediaQuery.of(context, nullOk: true);
     return PlatformApp(
       title: 'Flutter Demo',
-      locale: DevicePreview.of(context).locale, // <--
+      locale: DevicePreview.of(context)?.locale, // <--
       builder: DevicePreview.appBuilder, // <--
       supportedLocales: const [
         Locale("en"),
         Locale("fr", "FR"),
-        Locale.fromSubtags(
-          languageCode: "zh",
-          scriptCode: "Hans",
-          countryCode: "CN",
-        ),
       ],
       localizationsDelegates: [
         GlobalMaterialLocalizations.delegate,
@@ -47,11 +32,15 @@ class MyApp extends StatelessWidget {
       android: (_) => MaterialAppData(
         theme: ThemeData.light(),
         darkTheme: ThemeData.dark(),
-        themeMode: mediaQuery.platformBrightness == Brightness.dark
+        themeMode: mediaQuery?.platformBrightness == Brightness.dark
             ? ThemeMode.dark
             : ThemeMode.light,
       ),
-      ios: (_) => CupertinoAppData(theme: CupertinoThemeData()),
+      ios: (_) => CupertinoAppData(
+        theme: CupertinoThemeData(
+          brightness: mediaQuery.platformBrightness,
+        ),
+      ),
       home: MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
@@ -67,14 +56,6 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      _counter++;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
     final mediaQuery = MediaQuery.of(context);
@@ -90,15 +71,28 @@ class _MyHomePageState extends State<MyHomePage> {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: <Widget>[
-            // If not precised `textScaleFactor`
-            // is inherited from mediaQuery
             PlatformText(
               'Hello, it is ' + dateFormat.format(DateTime.now()),
               style: theme.textTheme.headline,
             ),
-            PlatformText(
-              '$_counter',
-              style: Theme.of(context).textTheme.display1,
+            PlatformButton(
+              child: PlatformText("Open"),
+              onPressed: () {
+                showPlatformDialog(
+                  context: context,
+                  builder: (context) => Container(
+                    height: 300,
+                    width: 300,
+                    padding: const EdgeInsets.all(20),
+                    child: PlatformButton(
+                      child: PlatformText("Close"),
+                      onPressed: () {
+                        Navigator.pop(context);
+                      },
+                    ),
+                  ),
+                );
+              },
             ),
           ],
         ),